refactor(web): extract the comment sidebar into the file-viewer slice
Moves CommentSidePanel (collapsed rail + expanded list, drag-reorder,
selection, and the inline new-comment composer) and its CommentSideDock
shell out of FileViewer.tsx and into the file-viewer vertical slice, per
ADR 0002. The attachment-thumbnail URL now goes through the slice's
existing pure fileRawUrl rule instead of importing providers/registry
directly, keeping the component transport-free.
