+++ apps/web/src/components/FileViewer.tsx
@@ -269,6 +269,8 @@ import {
   CommentPreviewOverlays,
   ReactComponentViewer,
   InspectPanel,
+  CommentSidePanel,
+  CommentSideDock,
 } from '../features/file-viewer';
 import type {
   InspectOverrideMap,
@@ -278,7 +280,6 @@ import type {
   PreviewScaleOptions,
   DeployProviderOption,
   ManualEditPendingStyleSave,
-  CommentSideDropEdge,
   PreviewOverlayTransform,
   BoardTool,
   StrokePoint,
@@ -301,6 +302,7 @@ export {
   appendSavedPreviewCommentOrder,
   SvgViewer,
   CommentTargetOverlay,
+  CommentSidePanel,
 } from '../features/file-viewer';
 export type { InspectOverrideEntry, InspectOverrideMap, ManualEditPendingStyleSave } from '../features/file-viewer';
 
@@ -2693,449 +2695,9 @@ function FileVersionManagerModal({
 }
 
 
-function commentDisplayLabel(comment: PreviewComment, t: TranslateFn): string {
-  if (comment.elementId.startsWith('pin-')) return t('chat.comments.pin');
-  const label = String(comment.label || '').trim().toLowerCase();
-  const htmlHint = String(comment.htmlHint || '').trim().toLowerCase();
-  const elementId = String(comment.elementId || '').trim().toLowerCase();
-  const source = `${label} ${htmlHint} ${elementId}`;
-  if (/\b(?:img|picture|video|canvas|svg)\b/.test(source)) return t('chat.comments.targetImage');
-  if (/\b(?:button|input|textarea|select|label)\b/.test(source)) return t('chat.comments.targetControl');
-  if (/^<a\b/.test(htmlHint)) return t('chat.comments.targetLink');
-  if (/\b(?:h1|h2|h3|h4|h5|h6|p|span|strong|em|small|li|dt|dd)\b/.test(source)) return t('chat.comments.targetText');
-  if (/\b(?:section|main|header|footer|nav|article|aside)\b/.test(source)) return t('chat.comments.targetSection');
-  if (label.endsWith('.html') || elementId.startsWith('file-comment-')) return t('chat.comments.targetPage');
-  if (comment.text.trim()) return t('chat.comments.targetText');
-  return t('chat.comments.targetArea');
-}
-
-export function CommentSidePanel({
-  comments,
-  projectId,
-  selectedIds,
-  activeCommentId,
-  collapsed,
-  onCollapsedChange,
-  onToggleSelect,
-  onSelectAll,
-  onClearSelection,
-  onReorder,
-  onReply,
-  onSendSelected,
-  onCreateComment,
-  sending,
-  queueOnSend = false,
-  sendDisabled = false,
-  renderCreateForm = true,
-  t,
-  composer,
-}: {
-  comments: PreviewComment[];
-  projectId?: string;
-  selectedIds: Set<string>;
-  activeCommentId: string | null;
-  collapsed: boolean;
-  onCollapsedChange: (collapsed: boolean) => void;
-  onToggleSelect: (commentId: string) => void;
-  onSelectAll: () => void;
-  onClearSelection: () => void;
-  onReorder?: (orderedIds: string[]) => void;
-  onReply: (comment: PreviewComment) => void;
-  onSendSelected: () => void | Promise<void>;
-  onCreateComment?: (note: string) => boolean | Promise<boolean>;
-  sending: boolean;
-  queueOnSend?: boolean;
-  sendDisabled?: boolean;
-  renderCreateForm?: boolean;
-  t: TranslateFn;
-  composer?: ReactNode;
-}) {
-  const [newCommentDraft, setNewCommentDraft] = useState('');
-  const [dragState, setDragState] = useState<CommentSideDragState | null>(null);
-  const sorted = comments;
-  const visibleSelectedIds = new Set(comments.filter((comment) => selectedIds.has(comment.id)).map((comment) => comment.id));
-  const selectedCount = visibleSelectedIds.size;
-  const allSelected = comments.length > 0 && selectedCount === comments.length;
-  const commentsLabel = t('chat.tabComments');
-  const canCreateComment = Boolean(onCreateComment) && newCommentDraft.trim().length > 0 && !sending && !sendDisabled;
-  const canReorder = Boolean(onReorder && sorted.length > 1);
-  const collapsedRailRef = useRef<HTMLButtonElement | null>(null);
-  const expandedToggleRef = useRef<HTMLButtonElement | null>(null);
-  const pendingToggleFocusRef = useRef<'collapsed' | 'expanded' | null>(null);
-  const panelId = useId();
-  const handleDragStart = (event: ReactDragEvent<HTMLButtonElement>, comment: PreviewComment) => {
-    if (!canReorder) return;
-    event.dataTransfer.effectAllowed = 'move';
-    event.dataTransfer.setData(COMMENT_SIDE_DRAG_MIME, comment.id);
-    event.dataTransfer.setData('text/plain', comment.id);
-    setDragState({ draggingId: comment.id, overId: comment.id, edge: null });
-  };
-  const handleDragOver = (event: ReactDragEvent<HTMLDivElement>, targetId: string) => {
-    if (!canReorder) return;
-    const draggingId = dragState?.draggingId || event.dataTransfer.getData(COMMENT_SIDE_DRAG_MIME);
-    if (!draggingId) return;
-    event.preventDefault();
-    event.dataTransfer.dropEffect = 'move';
-    if (draggingId === targetId) {
-      if (dragState?.overId !== targetId || dragState.edge !== null) {
-        setDragState({ draggingId, overId: targetId, edge: null });
-      }
-      return;
-    }
-    const edge = commentSideDropEdgeForEvent(event);
-    if (
-      dragState?.draggingId !== draggingId ||
-      dragState.overId !== targetId ||
-      dragState.edge !== edge
-    ) {
-      setDragState({ draggingId, overId: targetId, edge });
-    }
-  };
-  const handleDrop = (event: ReactDragEvent<HTMLDivElement>, targetId: string) => {
-    if (!canReorder) return;
-    event.preventDefault();
-    const draggingId =
-      dragState?.draggingId ||
-      event.dataTransfer.getData(COMMENT_SIDE_DRAG_MIME) ||
-      event.dataTransfer.getData('text/plain');
-    if (!draggingId || draggingId === targetId) {
-      setDragState(null);
-      return;
-    }
-    const edge = dragState?.overId === targetId && dragState.edge
-      ? dragState.edge
-      : commentSideDropEdgeForEvent(event);
-    const nextIds = reorderPreviewCommentIds(sorted, draggingId, targetId, edge);
-    if (nextIds.join('\0') !== sorted.map((comment) => comment.id).join('\0')) {
-      onReorder?.(nextIds);
-    }
-    setDragState(null);
-  };
-  const submitNewComment = async () => {
-    if (!onCreateComment || !newCommentDraft.trim()) return;
-    const saved = await onCreateComment(newCommentDraft.trim());
-    if (saved) setNewCommentDraft('');
-  };
-
-  useEffect(() => {
-    const target =
-      pendingToggleFocusRef.current === 'collapsed'
-        ? collapsedRailRef.current
-        : pendingToggleFocusRef.current === 'expanded'
-          ? expandedToggleRef.current
-          : null;
-    if (!target) return;
-    pendingToggleFocusRef.current = null;
-    target.focus();
-  }, [collapsed]);
-
-  const handleCollapsedChange = (
-    nextCollapsed: boolean,
-    nextFocusTarget: 'collapsed' | 'expanded',
-  ) => {
-    pendingToggleFocusRef.current = nextFocusTarget;
-    onCollapsedChange(nextCollapsed);
-  };
-
-  if (collapsed) {
-    return (
-      <button
-        ref={collapsedRailRef}
-        type="button"
-        className="comment-side-rail"
-        data-testid="comment-side-collapsed-rail"
-        aria-label={t('preview.showSidebar', { label: commentsLabel })}
-        aria-expanded={false}
-        title={t('preview.showSidebar', { label: commentsLabel })}
-        onClick={() => handleCollapsedChange(false, 'expanded')}
-      >
-        <RemixIcon name="message-3-line" size={15} />
-        <span>{commentsLabel}</span>
-        {comments.length > 0 ? <strong>{comments.length}</strong> : null}
-      </button>
-    );
-  }
-
-  return (
-    <aside id={panelId} className="comment-side-panel" data-testid="comment-side-panel" aria-label={commentsLabel}>
-      <div className="comment-side-header">
-        <div className="comment-side-title">
-          <RemixIcon name="message-3-line" size={15} />
-          <span>{commentsLabel}</span>
-        </div>
-        <div className="comment-side-header-actions">
-          {comments.length > 0 ? (
-            <button
-              type="button"
-              className="comment-side-select-all"
-              disabled={allSelected}
-              onClick={onSelectAll}
-            >
-              {t('chat.comments.selectAll')}
-            </button>
-          ) : null}
-          <button
-            ref={expandedToggleRef}
-            type="button"
-            className="comment-side-collapse"
-            aria-label={t('preview.hideSidebar', { label: commentsLabel })}
-            aria-controls={panelId}
-            aria-expanded={true}
-            title={t('preview.hideSidebar', { label: commentsLabel })}
-            onClick={() => handleCollapsedChange(true, 'collapsed')}
-          >
-            <Icon name="chevron-right" size={14} />
-          </button>
-        </div>
-      </div>
-      <div
-        className="comment-side-list"
-        onDragLeave={(event) => {
-          const related = event.relatedTarget;
-          if (related instanceof Node && event.currentTarget.contains(related)) return;
-          setDragState(null);
-        }}
-      >
-        {sorted.length === 0 ? (
-          <div className="comment-side-empty">
-            {t('chat.comments.emptySaved')}
-          </div>
-        ) : sorted.map((comment, index) => {
-          const selected = visibleSelectedIds.has(comment.id);
-          const active = comment.id === activeCommentId;
-          const isDragging = dragState?.draggingId === comment.id;
-          const dropClass = dragState?.overId === comment.id &&
-            dragState.draggingId !== comment.id &&
-            dragState.edge
-            ? ` comment-side-item-drop-${dragState.edge}`
-            : '';
-          return (
-            <div
-              key={comment.id}
-              className={`comment-side-item${selected ? ' selected' : ''}${active ? ' active' : ''}${isDragging ? ' dragging' : ''}${dropClass}`}
-              data-testid="comment-side-item"
-              data-comment-id={comment.id}
-              aria-current={active ? 'true' : undefined}
-              role="button"
-              tabIndex={0}
-              onDragOver={(event) => handleDragOver(event, comment.id)}
-              onDrop={(event) => handleDrop(event, comment.id)}
-              onClick={() => onReply(comment)}
-              onKeyDown={(event) => {
-                if (event.key !== 'Enter' && event.key !== ' ') return;
-                event.preventDefault();
-                onReply(comment);
-              }}
-            >
-              <div className="comment-side-item-head">
-                <button
-                  type="button"
-                  className="comment-side-drag-handle"
-                  title={t('chat.queuedReorder')}
-                  aria-label={t('chat.queuedReorder')}
-                  draggable={canReorder}
-                  disabled={!canReorder}
-                  onClick={(event) => event.stopPropagation()}
-                  onDragStart={(event) => handleDragStart(event, comment)}
-                  onDragEnd={() => setDragState(null)}
-                >
-                  <Icon name="grip-vertical" size={13} />
-                </button>
-                <span className="comment-side-author">
-                  <strong>{`${index + 1}. ${commentDisplayLabel(comment, t)}`}</strong>
-                </span>
-                <span className="comment-side-time">{formatCommentTime(commentActivityAt(comment), t)}</span>
-                <button
-                  type="button"
-                  className={`comment-side-check${selected ? ' checked' : ''}`}
-                  aria-label={selected ? t('chat.comments.deselect') : t('chat.comments.select')}
-                  aria-pressed={selected}
-                  onClick={(event) => {
-                    event.stopPropagation();
-                    onToggleSelect(comment.id);
-                  }}
-                >
-                  {selected ? <Icon name="check" size={11} /> : null}
-                </button>
-              </div>
-              <div className="comment-side-body">{comment.note}</div>
-              {projectId && comment.attachments && comment.attachments.length > 0 ? (
-                <div className="comment-side-attachments">
-                  {comment.attachments.map((attachment) => {
-                    const url = projectRawUrl(projectId, attachment.path);
-                    return (
-                      <a
-                        key={attachment.path}
-                        className="comment-side-attachment"
-                        data-testid="comment-side-attachment"
-                        href={url}
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        aria-label={attachment.name}
-                        title={attachment.name}
-                        onClick={(event) => event.stopPropagation()}
-                      >
-                        <img src={url} alt={attachment.name} />
-                      </a>
-                    );
-                  })}
-                </div>
-              ) : null}
-            </div>
-          );
-        })}
-      </div>
-      {selectedCount > 0 ? (
-        <div className="comment-side-selectbar" data-testid="comment-side-selectbar">
-          <span className="comment-side-selectcount">{t('chat.comments.nSelected', { n: selectedCount })}</span>
-          <Button variant="ghost" onClick={onClearSelection}>
-            {t('chat.comments.clear')}
-          </Button>
-          <Button
-            variant="primary"
-            data-testid="comment-side-send-claude"
-            disabled={sending || sendDisabled}
-            onClick={() => void onSendSelected()}
-          >
-            {sending
-              ? t('chat.comments.sending')
-              : queueOnSend
-                ? t('chat.annotationQueue')
-                : t('chat.comments.sendToChat')}
-          </Button>
-        </div>
-      ) : null}
-      {composer ? <div className="comment-side-composer">{composer}</div> : null}
-      {renderCreateForm && onCreateComment ? (
-        <form
-          className="comment-side-new-comment composer"
-          onSubmit={(event) => {
-            event.preventDefault();
-            void submitNewComment();
-          }}
-        >
-          <div className="composer-shell comment-side-new-comment-shell">
-            <div className="composer-input-wrap">
-              <div className="composer-textarea-layer">
-                <textarea
-                  value={newCommentDraft}
-                  placeholder={t('chat.comments.placeholder')}
-                  aria-label={t('chat.comments.placeholder')}
-                  onChange={(event) => setNewCommentDraft(event.target.value)}
-                  onKeyDown={(event) => {
-                    if ((event.metaKey || event.ctrlKey) && event.key === 'Enter') {
-                      event.preventDefault();
-                      void submitNewComment();
-                    }
-                  }}
-                />
-              </div>
-            </div>
-            <div className="composer-row comment-side-new-comment-actions">
-              <button
-                type="button"
-                className="icon-btn"
-                title={t('chat.attachTitle')}
-                aria-label={t('chat.attachAria')}
-                disabled
-              >
-                <Icon name="attach" size={15} />
-              </button>
-              <span className="composer-spacer" />
-              <button
-                type="submit"
-                className={`composer-send${sending ? ' is-sending' : ''}`}
-                disabled={!canCreateComment}
-              >
-                <Icon name="send" size={13} />
-                <span>{sending ? t('chat.comments.sending') : t('chat.send')}</span>
-              </button>
-            </div>
-          </div>
-        </form>
-      ) : null}
-    </aside>
-  );
-}
-
-const COMMENT_SIDE_DRAG_MIME = 'application/x-open-design-preview-comment';
-
-interface CommentSideDragState {
-  draggingId: string;
-  overId: string | null;
-  edge: CommentSideDropEdge | null;
-}
-
-function CommentSideDock({
-  comments,
-  projectId,
-  selectedIds,
-  activeCommentId,
-  collapsed,
-  onCollapsedChange,
-  onToggleSelect,
-  onSelectAll,
-  onClearSelection,
-  onReorder,
-  onReply,
-  onSendSelected,
-  onCreateComment,
-  sending,
-  queueOnSend = false,
-  sendDisabled = false,
-  renderCreateForm = true,
-  t,
-  composer,
-}: {
-  comments: PreviewComment[];
-  projectId?: string;
-  selectedIds: Set<string>;
-  activeCommentId: string | null;
-  collapsed: boolean;
-  onCollapsedChange: (collapsed: boolean) => void;
-  onToggleSelect: (commentId: string) => void;
-  onSelectAll: () => void;
-  onClearSelection: () => void;
-  onReorder?: (orderedIds: string[]) => void;
-  onReply: (comment: PreviewComment) => void;
-  onSendSelected: () => void | Promise<void>;
-  onCreateComment?: (note: string) => boolean | Promise<boolean>;
-  sending: boolean;
-  queueOnSend?: boolean;
-  sendDisabled?: boolean;
-  renderCreateForm?: boolean;
-  t: TranslateFn;
-  composer?: ReactNode;
-}) {
-  return (
-    <div
-      className={`comment-side-dock${collapsed ? ' collapsed' : ''}`}
-      data-testid="comment-side-dock"
-    >
-      <CommentSidePanel
-        comments={comments}
-        projectId={projectId}
-        selectedIds={selectedIds}
-        activeCommentId={activeCommentId}
-        collapsed={collapsed}
-        onCollapsedChange={onCollapsedChange}
-        onToggleSelect={onToggleSelect}
-        onSelectAll={onSelectAll}
-        onClearSelection={onClearSelection}
-        onReorder={onReorder}
-        onReply={onReply}
-        onSendSelected={onSendSelected}
-        onCreateComment={onCreateComment}
-        sending={sending}
-        queueOnSend={queueOnSend}
-        sendDisabled={sendDisabled}
-        renderCreateForm={renderCreateForm}
-        t={t}
-        composer={composer}
-      />
-    </div>
-  );
-}
+// CommentSidePanel/CommentSideDock (comment sidebar, drag-reorder,
+// composer) now live in the file-viewer slice; imported from the slice
+// barrel above (ADR 0002).
 
 // InspectPanel now lives in the file-viewer slice; imported from the
 // slice barrel above (ADR 0002).
