+++ apps/web/src/features/file-viewer/components/CommentSideDock.tsx
@@ -1,0 +1,77 @@
+// Thin dock shell around CommentSidePanel: applies the collapsed/expanded
+// layout class the preview chrome positions against.
+import type { ReactNode } from 'react';
+import type { PreviewComment } from '../../../types';
+import type { TranslateFn } from '../types';
+import { CommentSidePanel } from './CommentSidePanel';
+
+export function CommentSideDock({
+  comments,
+  projectId,
+  selectedIds,
+  activeCommentId,
+  collapsed,
+  onCollapsedChange,
+  onToggleSelect,
+  onSelectAll,
+  onClearSelection,
+  onReorder,
+  onReply,
+  onSendSelected,
+  onCreateComment,
+  sending,
+  queueOnSend = false,
+  sendDisabled = false,
+  renderCreateForm = true,
+  t,
+  composer,
+}: {
+  comments: PreviewComment[];
+  projectId?: string;
+  selectedIds: Set<string>;
+  activeCommentId: string | null;
+  collapsed: boolean;
+  onCollapsedChange: (collapsed: boolean) => void;
+  onToggleSelect: (commentId: string) => void;
+  onSelectAll: () => void;
+  onClearSelection: () => void;
+  onReorder?: (orderedIds: string[]) => void;
+  onReply: (comment: PreviewComment) => void;
+  onSendSelected: () => void | Promise<void>;
+  onCreateComment?: (note: string) => boolean | Promise<boolean>;
+  sending: boolean;
+  queueOnSend?: boolean;
+  sendDisabled?: boolean;
+  renderCreateForm?: boolean;
+  t: TranslateFn;
+  composer?: ReactNode;
+}) {
+  return (
+    <div
+      className={`comment-side-dock${collapsed ? ' collapsed' : ''}`}
+      data-testid="comment-side-dock"
+    >
+      <CommentSidePanel
+        comments={comments}
+        projectId={projectId}
+        selectedIds={selectedIds}
+        activeCommentId={activeCommentId}
+        collapsed={collapsed}
+        onCollapsedChange={onCollapsedChange}
+        onToggleSelect={onToggleSelect}
+        onSelectAll={onSelectAll}
+        onClearSelection={onClearSelection}
+        onReorder={onReorder}
+        onReply={onReply}
+        onSendSelected={onSendSelected}
+        onCreateComment={onCreateComment}
+        sending={sending}
+        queueOnSend={queueOnSend}
+        sendDisabled={sendDisabled}
+        renderCreateForm={renderCreateForm}
+        t={t}
+        composer={composer}
+      />
+    </div>
+  );
+}
